+++ sdk/formrecognizer/azure-ai-formrecognizer/azure/ai/formrecognizer/aio/_form_recognizer_client_async.py
@@ -90,10 +90,14 @@ class FormRecognizerClient(FormRecognizerClientBaseAsync):
         :keyword int polling_interval: Waiting time between two polls for LRO operations
             if no Retry-After header is present. Defaults to 5 seconds.
         :keyword str continuation_token: A continuation token to restart a poller from a saved state.
+        :keyword str locale: Locale of the receipt. Defaults to en-US. Supported locales include:
+            en-US, en-AU, en-CA, en-GB, and en-IN.
         :return: An instance of an AsyncLROPoller. Call `result()` on the poller
             object to return a list[:class:`~azure.ai.formrecognizer.RecognizedForm`].
         :rtype: ~azure.core.polling.AsyncLROPoller[list[~azure.ai.formrecognizer.RecognizedForm]]
         :raises ~azure.core.exceptions.HttpResponseError:
+        .. versionadded:: v2.1-preview
+            The *locale* keyword argument
 
         .. admonition:: Example:
 
@@ -145,10 +149,14 @@ class FormRecognizerClient(FormRecognizerClientBaseAsync):
         :keyword int polling_interval: Waiting time between two polls for LRO operations
             if no Retry-After header is present. Defaults to 5 seconds.
         :keyword str continuation_token: A continuation token to restart a poller from a saved state.
+        :keyword str locale: Locale of the receipt. Defaults to en-US. Supported locales include:
+            en-US, en-AU, en-CA, en-GB, and en-IN.
         :return: An instance of an AsyncLROPoller. Call `result()` on the poller
             object to return a list[:class:`~azure.ai.formrecognizer.RecognizedForm`].
         :rtype: ~azure.core.polling.AsyncLROPoller[list[~azure.ai.formrecognizer.RecognizedForm]]
         :raises ~azure.core.exceptions.HttpResponseError:
+        .. versionadded:: v2.1-preview
+            The *locale* keyword argument
 
         .. admonition:: Example:
 
@@ -394,6 +402,7 @@ class FormRecognizerClient(FormRecognizerClientBaseAsync):
 
         polling_interval = kwargs.pop("polling_interval", self._client._config.polling_interval)
         content_type = kwargs.pop("content_type", None)
+        continuation_token = kwargs.pop("continuation_token", None)
         if content_type == "application/json":
             raise TypeError("Call begin_recognize_custom_forms_from_url() to analyze a document from a URL.")
 
@@ -417,6 +426,7 @@ class FormRecognizerClient(FormRecognizerClientBaseAsync):
                 lro_algorithms=[AnalyzePolling()],
                 **kwargs
             ),
+            continuation_token=continuation_token,
             **kwargs
         )
 
@@ -448,7 +458,7 @@ class FormRecognizerClient(FormRecognizerClientBaseAsync):
         if not model_id:
             raise ValueError("model_id cannot be None or empty.")
         polling_interval = kwargs.pop("polling_interval", self._client._config.polling_interval)
-
+        continuation_token = kwargs.pop("continuation_token", None)
         include_field_elements = kwargs.pop("include_field_elements", False)
 
         def analyze_callback(raw_response, _, headers):  # pylint: disable=unused-argument
@@ -465,6 +475,7 @@ class FormRecognizerClient(FormRecognizerClientBaseAsync):
                 lro_algorithms=[AnalyzePolling()],
                 **kwargs
             ),
+            continuation_token=continuation_token,
             **kwargs
         )