+++ sdk/servicebus/azure-servicebus/azure/servicebus/_common/_configuration.py
@@ -20,7 +20,9 @@ class Configuration(object):  # pylint:disable=too-many-instance-attributes
             if self.http_proxy
             else kwargs.get("transport_type", TransportType.Amqp)
         )
+        # The following configs are not public, for internal usage only
         self.auth_timeout = kwargs.get("auth_timeout", 60)  # type: int
         self.encoding = kwargs.get("encoding", "UTF-8")
         self.auto_reconnect = kwargs.get("auto_reconnect", True)
         self.keep_alive = kwargs.get("keep_alive", 30)
+        self.timeout = kwargs.get("timeout", 60)  # type: float